+++ DatadogSessionReplay/SRSnapshotTests/SRSnapshotTests/Utils/SnapshotTestCase.swift
@@ -15,6 +15,10 @@ private var defaultPrivacyLevel: SessionReplayPrivacyLevel {
     return SessionReplay.Configuration(replaySampleRate: 100).defaultPrivacyLevel
 }
 
+private var defaultImagePrivacyLevel: ImagePrivacyLevel {
+    return .maskNonBundledOnly
+}
+
 internal class SnapshotTestCase: XCTestCase {
     private var app: AppDelegate { UIApplication.shared.delegate as! AppDelegate }
 
@@ -39,6 +43,7 @@ internal class SnapshotTestCase: XCTestCase {
     func takeSnapshotFor(
         _ fixture: Fixture,
         with privacyModes: [SessionReplayPrivacyLevel] = [defaultPrivacyLevel],
+        imagePrivacyLevel: ImagePrivacyLevel = defaultImagePrivacyLevel,
         shouldRecord: Bool,
         folderPath: String,
         fileNamePrefix: String? = nil,
@@ -50,7 +55,7 @@ internal class SnapshotTestCase: XCTestCase {
         wait(seconds: 0.2)
 
         try forPrivacyModes(privacyModes) { privacyMode in
-            let image = try takeSnapshot(with: privacyMode)
+            let image = try takeSnapshot(with: privacyMode, imagePrivacyLevel: imagePrivacyLevel)
             let fileNameSuffix = fileNamePrefix == nil ? "-\(privacyMode)-privacy" : "-\(fileNamePrefix!)-\(privacyMode)-privacy"
             let snapshotLocation: ImageLocation = .folder(named: folderPath, fileNameSuffix: fileNameSuffix, file: file, function: function)
 
@@ -115,7 +120,10 @@ internal class SnapshotTestCase: XCTestCase {
     }
 
     /// Captures side-by-side snapshot of the app UI and recorded wireframes.
-    func takeSnapshot(with privacyLevel: SessionReplayPrivacyLevel = defaultPrivacyLevel) throws -> UIImage {
+    func takeSnapshot(
+        with privacyLevel: SessionReplayPrivacyLevel = defaultPrivacyLevel,
+        imagePrivacyLevel: ImagePrivacyLevel = defaultImagePrivacyLevel
+    ) throws -> UIImage {
         let expectWireframes = self.expectation(description: "Wait for wireframes")
         let expectResources = self.expectation(description: "Wait for resources")
 
@@ -154,7 +162,16 @@ internal class SnapshotTestCase: XCTestCase {
 
         // Capture next record with mock RUM Context
         try recorder.captureNextRecord(
-            .init(privacy: privacyLevel, textAndInputPrivacy: .maskSensitiveInputs, touchPrivacy: .show, applicationID: "", sessionID: "", viewID: "", viewServerTimeOffset: 0)
+            .init(
+                privacy: privacyLevel,
+                textAndInputPrivacy: .maskSensitiveInputs,
+                imagePrivacy: imagePrivacyLevel,
+                touchPrivacy: .show,
+                applicationID: "",
+                sessionID: "",
+                viewID: "",
+                viewServerTimeOffset: 0
+            )
         )
 
         waitForExpectations(timeout: 30) // very pessimistic timeout to mitigate CI lags